# src/swagger_doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Starnet API",
    "description": ""
  },
  "servers": [
    {
      "url": "https://dev.starnet.ai/api.staging",
      "description": ""
    }
  ],
  "basePath": "",
...
  "paths": {
...
    "/upload/{resource}/{id}": {
      "post": {
        "tags": [
          "Utilities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file",
            "in": "formData",
            "required": true,
            "description": "Select a file to upload",
            "schema": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/upload/posts/{campaign_id}/{user_id}": {
      "post": {
        "tags": [
          "Utilities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photos",
            "in": "formData",
            "required": false,
            "description": "Select up to 10 images to upload",
            "schema": {
              "type": "file"
            }
          },
          {
            "name": "title",
            "in": "formData",
            "required": false,
            "description": "The title of the post",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content",
            "in": "formData",
            "required": false,
            "description": "The content of the post",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}